==> 3_creacion_de_tablas.ipynb <==
----------

# Creación de tablas a partir de las columnas con datos anidados

6- Algunos campos, como belongs_to_collection, production_companies y otros (ver diccionario de datos) están anidados, esto es o bien tienen un diccionario o una lista como valores en cada fila, ¡deberán desanidarlos y unirlos al dataset de nuevo para hacer alguna de las consultas de la API! O bien buscar la manera de acceder a esos datos sin desanidarlos..

Como se mencionó anteriormente, las columnas que poseen datos anidados son: 'belongs_to_collection', 'genres', 'production_companies','production_countries' y 'spoken_languages'

Para la función peliculas_idioma nos piden devolver la cantidad de películas producidas en ese idioma. Por lo que no utilizaremos la columna spoken_language y será eliminada.

(x) Para la función peliculas_duracion utilizaremos la columna runtime

(x) Para la funcion franquicia deberemos desanidar la columna belongs_to_collection

(x) Para la función pais deberemos desanidar la columna production_countries

(x) Para la función productoras_exitosas deberemos desanidar la columna production_companies

() Para la función get_director deberemos usar el dataset credits. @@64♠z4♠.4♠46♦6♦6♦6♦621♣:Õ☻

(x)La columna genres la utilizaremos para el sistema de recomendacion por lo que la deberemos desanidar.

índice de tablas:
- Tabla Collections
- Tablas movies_production_companies y production_companies
- Tablas movies_production_countries y production_countries
- Tabla genres
- Tabla original_language

----

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import ast
import seaborn as sns

# Establecer la opción para mostrar todas las columnas
pd.set_option('display.max_columns', None)

movies = pd.read_csv('data/movies_2.csv',low_memory=False)


movies = movies.drop(columns=['spoken_languages'])

movies.columns

movies.head(3)

movies.info()

------

# Cración de tablas


---------

# Tabla Collections

Creación de la tabla 'collections' que utilizaremos en la función franquicia()

# cremaos la columna 'name_collecction', extraemos sus valores de la columna belongs_to_collection y luego eliminamos esta última.

# # Definimos una expresión regular para extraer el id del campo "id" de la columna belongs_to_collection
regex_name = r"'name':\s*'([^']+)"
regex_id = r"'id':\s*([\d]+)"


# # Utilizamos la función "str.extract" para extraer el id y guardarlo en una nueva columna 'name_collection'
movies['name_collection'] = movies['belongs_to_collection'].str.extract(regex_name)
movies['id_collection'] = movies['belongs_to_collection'].str.extract(regex_id)

# Eliminamos la columna 'belongs_to_collection'
movies.drop(['belongs_to_collection'], axis= 'columns', inplace=True)

movies.columns

movies.head()

collections = movies[['id_movie','name_collection','revenue']]

collections.info()

(
    collections
    .isnull()
    .sum()
)

collections[collections['name_collection'].isnull()]

Existen peliculas que no pertenecen a ninguna franquicia por lo que las eliminaremos de la tabla 'collections'

collections = collections.dropna()

collections.head(2)

# Inicial de las palabras con mayusculas
collections['name_collection'] = collections['name_collection'].apply(lambda x: x.title().strip())

# crear el csv 
collections.to_csv("data/collections.csv",index=False)

------------

# Tablas movies_production_companies y production_companies

Creacion de las tablas movies_production_companies y production_companies utilizaremos en la función productoras_exitosas()

movies.columns

# Creamos la tabla movies_production_companies a partir de las columnas 'id_movie','production_companies' y 'revenue'
# del df movies, que serán necesarias para crear la funcion productoras_exitosas

movies_production_companies = movies[['id_movie','production_companies','revenue']].copy()

movies_production_companies.shape

movies_production_companies.tail(10)

movies_production_companies[movies_production_companies.production_companies == '[]'].shape

Podemos ver que hay peliculas que no tienen especificadas la compañia productora, no aportando informacion a este dataset, por lo que las eliminaremos.

movies_production_companies = movies_production_companies[movies_production_companies.production_companies != '[]']

movies_production_companies.shape

# Convertimos las cadenas de texto que representan listas en "production_companies" en listas de python.
movies_production_companies['production_companies'] = movies_production_companies['production_companies'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista production_companies
movies_production_companies = movies_production_companies.explode('production_companies')

# Extraemos el valor de id_production_companies de cada elemento en la lista production_companies
movies_production_companies['id_production_companies'] = movies_production_companies['production_companies'].apply(lambda x: x['id'])

# Extraemos el valor de id_production_companies de cada elemento en la lista production_companies
movies_production_companies['name_production_companies'] = movies_production_companies['production_companies'].apply(lambda x: x['name'])

movies_production_companies.columns

# ordenamos las columnas
movies_production_companies = movies_production_companies[['id_movie','id_production_companies','name_production_companies','revenue']]
movies_production_companies.head(3)

movies_production_companies.name_production_companies = movies_production_companies.name_production_companies.str.title()
movies_production_companies

# Creamos el csv
movies_production_companies.to_csv("data/movies_production_companies.csv",index=False)

-----

# Production_companies

## TABLA production_companies 

production_companies = movies['production_companies'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
production_companies = pd.DataFrame([dict(x) for x in production_companies.sum()])[['id', 'name']]
production_companies = production_companies.drop_duplicates()
production_companies.to_csv("data/production_companies.csv",index=False)

production_companies

# Eliminamos la columna production_companies de la tabla movies
movies = movies.drop('production_companies', axis=1)

--------

# Tablas movies_production_countries y production_countries

Creacion de las tablas movies_production_companies y production_companies utilizaremos en la función movies_production_countries()

### TABLA movies_production_countries

movies_production_countries = movies[['id_movie','production_countries']].copy()

# Reemplazamos los valores nulos por "[{'iso_3166_1': 'none', 'name': 'none'}]"
movies_production_countries['production_countries'] = movies_production_countries[['production_countries']].replace("[]", "[{'iso_3166_1': 'none', 'name': 'none'}]")  # ya esta arriba
movies_production_countries['production_countries'] = movies_production_countries['production_countries'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista production_countries
movies_production_countries = movies_production_countries.explode('production_countries')

# Extraer el valor de id_production_countries de cada elemento en la lista production_countries
movies_production_countries['id_production_countries'] = movies_production_countries['production_countries'].apply(lambda x: x['iso_3166_1'])

# Eliminar la columna genres
movies_production_countries = movies_production_countries.drop('production_countries', axis=1)

# Creamos el csv
movies_production_countries.to_csv("data/movies_production_countries.csv",index=False)

movies_production_countries

### Tabla production_countries 
#Esta tabla no se utilizara pero se crea para conservar el significado de los id_production_countries de la tabla movies_production_countries

production_countries = movies['production_countries'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
production_countries = pd.DataFrame([dict(x) for x in production_countries.sum()])[['iso_3166_1', 'name']]
production_countries = production_countries.drop_duplicates()

# Creamos el csv
production_countries.to_csv("data/production_countries.csv",index=False)

# Eliminamos la columna production_countries de la tabla movies
movies = movies.drop('production_countries', axis=1)

production_countries

movies.shape

Cabe notar que el dataset movies posee peliculas con 161 idiomas distintos. Se analiza en el archivo ..............AGREGAR NOMBREEEEEE

-----------

# Tabla genres

Creacion de la tabla genres que utilizaremos para el sistema de recomendación y sus tablas auxiliares.

movies.head()

### Tabla genres

genres = movies['genres'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
genres = pd.DataFrame([dict(x) for x in genres.sum()])[['id', 'name']]
genres = genres.drop_duplicates()
genres.to_csv("data/genres.csv",index=False)


### Tabla movies_genres
movies_genres = movies[['id_movie','genres']].copy()

# Generamos la categoria none para las peliculas que no tienen un genero especificado
movies_genres['genres'] = movies_genres[['genres']].replace("[]", "[{'id': 0, 'name': 'None'}]")  

# Convertimos los valores de movies_genres['genres'] (que parecen una lista) en un objeto tipo lista
movies_genres['genres'] = movies_genres['genres'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista genres
movies_genres = movies_genres.explode('genres')

# Extraer el valor de id_genre de cada elemento en la lista genres
movies_genres['id_genre'] = movies_genres['genres'].apply(lambda x: x['id'])


movies_genres

movies.head(2)

### Tabla m
ml_genres = movies[['title','genres','id_movie']].copy()

# Generamos la categoria none para las peliculas que no tienen un genero especificado
ml_genres['genres'] = ml_genres['genres'][ml_genres['genres'] != '[]']



ml_genres

ml_genres[ml_genres.genres.isna()]

ml_genres = ml_genres.dropna(subset='genres')
# Convertimos los valores de movies_genres['genres'] (que parecen una lista) en un objeto tipo lista
ml_genres['genres'] = ml_genres['genres'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista genres
ml_genres = ml_genres.explode('genres')

# Extraer el valor de id_genre de cada elemento en la lista genres
ml_genres['genres'] = ml_genres['genres'].apply(lambda x: x['name'])

ml_genres

ml_genres = ml_genres.groupby(['title','id_movie'],as_index=False)['genres'].agg(' '.join)

# Eliminar la columna genres de movies
movies = movies.drop('genres', axis=1)

# Eliminar la columna genres de movies_genres
movies_genres = movies_genres.drop('genres', axis=1)

# Creamos el csv
movies_genres.to_csv("data/movies_genres.csv",index=False)
ml_genres.to_csv("data/ml_genres.csv",index=False)


ml_genres

------

# Tabla original_language

Creación de la tabla 'original_language' que utilizaremos en la función peliculas_idioma()

original_language = movies[['id_movie','title','original_language']].copy(deep=False)

original_language

original_language.to_csv("data/original_language.csv",index=False)


-----------------

movies.to_csv("data/movies_3.csv",index=False)
