==> 7_etl_ML.ipynb <==
Importamos las librerias que vamos a utilizar.

import pandas as pd
import numpy as np
import ast

# Cargamos los datasets que vamos a utilizar.

credits = pd.read_csv('data/credits.csv',low_memory=False)

# Analisis de credits

credits.head(2)

credits.info()

Analisis de la columna cast

# cast
print (credits.iloc[0,0]) # ejemplo de los datos
print (type(credits.iloc[0,0])) # tipo de los datos

Analisis de la columna crew

# crew
print (credits.iloc[0,1]) # ejemplo de los datos
print (type(credits.iloc[0,1])) # tipo de los datos


-----------

# Extracción de los datos de la columna crew. Creacion de la tabla directors_crew.

print (ast.literal_eval(credits.iloc[0,1])[0].keys()) # Obtener las llaves del diccionario

credits

# Convertimos los str dentro de cast en estructuras de python correspondiente, en este caso en una lista
credits['crew'] = credits['crew'].apply(ast.literal_eval) 


# Explode('cast') desanida la columna cast y crea un nuevo df con filas duplicadas segun los elementos de la lista en la columna cast. 
# De este nuevo df solo seleccionamos las columnas que vamos a utilizar.
directors_crew = credits.explode('crew')['crew'].apply(pd.Series)[['id','name']] 

# Agrega la columna id_movie al df actors
directors_crew['id_movie'] = credits['id'] 

directors_crew

directors_crew.head(2)

Exploramos la tabla directors_crew

directors_crew.info()

directors_crew.max()

¿Hay valores nulos? ¿Cuántos?

(
    directors_crew
    .isnull()
    .sum()
)

directors_crew = directors_crew.dropna(subset='id')


Eliminamos las filas con valores nulos ya que no aportan información relevante.

Corregimos los tipo de dato de las columnas

directors_crew

directors_crew = directors_crew.astype({'id':'int64', 'name':'str'}, copy=False, errors='raise')

Existen filas duplicadas?

directors_crew[directors_crew.duplicated()].count()

directors_crew = directors_crew.drop_duplicates()

directors_crew.info()

directors_crew.to_csv("data/directors_crew.csv",index=False)

-------
